requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.0
jupyter-core==4.10.0
loguru==0.6.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
regex==2022.4.24
requests==2.27.1
six==1.16.0
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
